sysSentry: Fix two code review comments
driver inclusion
category: feature
bugzilla: https://gitee.com/openeuler/sysSentry/issues/ICV952
CVE: NA

----------------------------------------------------------------------

Fix two code review comments:
a. When a uint64_t type loop_cnt is passed as a parameter to an int type,
truncation occurs. To resolve this issue, the parameter type is changed
to uint64_t.

b. The loop counter loop_cnt is of type uint64_t. When it increments to
its maximum value and then adds 1, it overflows to 0, causing an infinite
loop that cannot be exited. To address this problem, a loop condition is
added to ensure that the loop continues only if loop_cnt does not overflow
to 0 after incrementing.

+++ src/sentryPlugins/soc_ring_sentry/tc_ring_one.c
@@ -641,7 +641,7 @@ static bool tc_ring_one_scan_bit(struct tc_ring_one_config *config, int scan_bit
     return true;
 }
 
-static bool tc_ring_one_test_issue(struct tc_ring_one_config *config, int scan_bit, int loop_cnt)
+static bool tc_ring_one_test_issue(struct tc_ring_one_config *config, int scan_bit, uint64_t loop_cnt)
 {
     bool ret;
